+++ litscan/__init__.py
@@ -6,12 +6,12 @@ Since: 1.0.0
 
 from env_dir_bootstrap import EnvDirBootstrap
 
-__version__ = "2.1.1"
+__version__ = "2.2.0"
 __app_name__ = "litscan"
 
 _bootstrapper = EnvDirBootstrap(
     env_var="LITSCAN_CONFIG_DIR",
-    resources=["logging.ini", "lit_ignore", ".litscanignore"],
+    resources=["logging.ini", "lit_ignore", ".litscanignore", "config.ini"],
     package="litscan",
 )